+++ src/layout/Page.tsx
@@ -1,7 +1,0 @@
-import styled from "../styled-components";
-
-/** Component that acts as a flex container and minimum height wrapper for the entire page */
-export const Page = styled.div`
-    flex: 1 1 auto;
-    min-height: 100vh;
-`;